AMC Ministries Solapur
Our Pastor
Suhas Bawadekar
Our Pastor - Rev.Suhas Bawadekar Started the Mission of preaching the gospel in the year 2012 with his team and his wife Sis.Flora Bawadekar, helping many Souls to enter the Heavenly Kingdom.He Started his mission both online and offline .He encourages All the members to testify, lead the worship and even Share the Word of God. We are thankful to god for leadership and service. we are grateful for all you do for our church. Thank you for being such a great example of living a life of faith. You're a blessing to our church. Thanks for all the ways you make a difference in our church and community.

Special Thanks to Pst Suhas, Sis.Flora Bawadekar for helping and encouraging me to make this website with the guidance of the Holy spirit

We Dedicate this website to Our Lord and savior Jesus Christ
Website Released on 24/10/2021

Developer : Jesse Obed
Helping Hand : Sis Flora Bawadekar, Anugrah Bawadekar
Pastor : Suhas Bawadekar